refactor: split ws/client.rs into domain handler modules
ws/client.rs was 1,490 lines handling 6 domains in one file with a
200-line match statement. Split into focused modules:

- ws/client.rs (305 lines) — connection lifecycle + event dispatch only
- ws/handlers/message.rs (235) — send, edit, delete, typing
- ws/handlers/reaction.rs (69) — add, remove
- ws/handlers/thread.rs (160) — thread message CRUD
- ws/handlers/voice.rs (325) — join, leave, answer, ICE, mute, deafen, screen, webcam
- ws/handlers/dm.rs (211) — DM send, edit, delete
- ws/handlers/request.rs (192) — request/response pattern

Pure refactor — no behavior changes. All handler signatures preserved.
Handler visibility is pub(in crate::ws) for module-internal access.
