+++ server-rs/src/ws/handlers/mod.rs
@@ -1,0 +1,13 @@
+mod dm;
+mod message;
+mod reaction;
+mod request;
+mod thread;
+mod voice;
+
+pub(in crate::ws) use dm::*;
+pub(in crate::ws) use message::*;
+pub(in crate::ws) use reaction::*;
+pub(in crate::ws) use request::*;
+pub(in crate::ws) use thread::*;
+pub(in crate::ws) use voice::*;
