+++ server-rs/src/ws/mod.rs
@@ -1,5 +1,6 @@
 pub mod events;
 pub mod hub;
+mod handlers;
 pub mod client;
 
 pub use hub::Hub;
